Refactor client handler unit tests into separate files
At over 1000 lines long, the single tests file for client handlers
has been growing incresingly unwieldy. Refactor into multiple
files to ease future development.
